Move MouseCommands into mouse package
