+++ libs/statemanager/src/main/kotlin/mono/state/MainStateManager.kt
@@ -24,8 +24,8 @@ import mono.shapebound.InteractionPoint
 import mono.shapesearcher.ShapeSearcher
 import mono.state.command.CommandEnvironment
 import mono.state.command.CommandType
-import mono.state.command.MouseCommand
-import mono.state.command.MouseCommandFactory
+import mono.state.command.mouse.MouseCommand
+import mono.state.command.mouse.MouseCommandFactory
 
 /**
  * A class which is connect components in the app.
