+++ libs/statemanager/src/main/kotlin/mono/state/command/mouse/LineInteractionMouseCommand.kt
@@ -1,4 +1,4 @@
-package mono.state.command
+package mono.state.command.mouse
 
 import mono.graphics.geo.DirectedPoint
 import mono.graphics.geo.MousePointer
@@ -7,6 +7,7 @@ import mono.shape.command.MoveLineAnchor
 import mono.shape.command.MoveLineEdge
 import mono.shape.shape.Line
 import mono.shapebound.LineInteractionPoint
+import mono.state.command.CommandEnvironment
 
 /**
  * A [MouseCommand] for moving anchor points or edges of Line shape.
