+++ libs/statemanager/src/main/kotlin/mono/state/command/mouse/ScaleShapeMouseCommand.kt
@@ -1,10 +1,11 @@
-package mono.state.command
+package mono.state.command.mouse
 
 import mono.graphics.geo.MousePointer
 import mono.graphics.geo.Point
 import mono.shape.command.ChangeBound
 import mono.shape.shape.AbstractShape
 import mono.shapebound.ScaleInteractionPoint
+import mono.state.command.CommandEnvironment
 
 /**
  * A [MouseCommand] for scaling shape.
